Make completed not deletable

+++ web/src/routes/fodnus.svelte
@@ -55,28 +55,46 @@
 			<thead>
 				<tr>
 					<td>
-						<div class="flex flex-row gap-2 place-items-center">
-							<button
-								on:click={() => {
-									const confirmed = confirm(
-										`Bekræft! - Er du sikker på du vil slette alle ${orders.length} ordre?`
-									)
-									if (confirmed) {
-										for (const order of orders) {
-											deleteOrder(order.id)
+						{#if orders.filter((val) => val.completed == false).length}
+							<div class="flex flex-row gap-2 place-items-center">
+								<button
+									on:click={() => {
+										const confirmed = confirm(
+											`Bekræft! - Er du sikker på du vil slette alle ${
+												orders.filter((val) => val.completed == false).length
+											} ordre?`
+										)
+										if (confirmed) {
+											for (const order of orders) {
+												if (!order.completed) {
+													deleteOrder(order.id)
+												}
+											}
 										}
-									}
-								}}
-								class="btn btn-error btn-outline"
-							>
-								<svg width="24" height="24" xmlns="http://www.w3.org/2000/svg" fill="currentColor"
-									><path
-										d="M19 24h-14c-1.104 0-2-.896-2-2v-17h-1v-2h6v-1.5c0-.827.673-1.5 1.5-1.5h5c.825 0 1.5.671 1.5 1.5v1.5h6v2h-1v17c0 1.104-.896 2-2 2zm0-19h-14v16.5c0 .276.224.5.5.5h13c.276 0 .5-.224.5-.5v-16.5zm-9 4c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm6 0c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm-2-7h-4v1h4v-1z"
-									/></svg
+									}}
+									class="btn btn-error btn-outline"
 								>
-							</button>
-							<h1 class="text-xl">{orders.length}</h1>
-						</div>
+									<svg width="24" height="24" xmlns="http://www.w3.org/2000/svg" fill="currentColor"
+										><path
+											d="M19 24h-14c-1.104 0-2-.896-2-2v-17h-1v-2h6v-1.5c0-.827.673-1.5 1.5-1.5h5c.825 0 1.5.671 1.5 1.5v1.5h6v2h-1v17c0 1.104-.896 2-2 2zm0-19h-14v16.5c0 .276.224.5.5.5h13c.276 0 .5-.224.5-.5v-16.5zm-9 4c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm6 0c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm-2-7h-4v1h4v-1z"
+										/></svg
+									>
+								</button>
+								<h1 class="text-xl">{orders.filter((val) => val.completed == false).length}</h1>
+							</div>
+						{:else}
+							<div class="flex flex-row gap-2 place-items-center">
+								<button class="btn btn-error btn-outline" disabled>
+									<svg width="24" height="24" xmlns="http://www.w3.org/2000/svg" fill="currentColor"
+										><path
+											d="M19 24h-14c-1.104 0-2-.896-2-2v-17h-1v-2h6v-1.5c0-.827.673-1.5 1.5-1.5h5c.825 0 1.5.671 1.5 1.5v1.5h6v2h-1v17c0 1.104-.896 2-2 2zm0-19h-14v16.5c0 .276.224.5.5.5h13c.276 0 .5-.224.5-.5v-16.5zm-9 4c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm6 0c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm-2-7h-4v1h4v-1z"
+										/></svg
+									>
+								</button>
+
+								<h1 class="text-xl">0</h1>
+							</div>
+						{/if}
 					</td>
 					<th>Tidspunkt</th>
 					<th>Beskrivelse</th>
@@ -89,21 +107,23 @@
 					{#each orders.sort((a, b) => a.completed - b.completed) as order, i}
 						<tr>
 							<td>
-								<div class="flex flex-row gap-2 place-items-center">
-									<button on:click={deleteOrder(order.id)} class="btn btn-square btn-outline">
-										<svg
-											width="24"
-											height="24"
-											xmlns="http://www.w3.org/2000/svg"
-											fill="currentColor"
-											><path
-												d="M19 24h-14c-1.104 0-2-.896-2-2v-17h-1v-2h6v-1.5c0-.827.673-1.5 1.5-1.5h5c.825 0 1.5.671 1.5 1.5v1.5h6v2h-1v17c0 1.104-.896 2-2 2zm0-19h-14v16.5c0 .276.224.5.5.5h13c.276 0 .5-.224.5-.5v-16.5zm-9 4c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm6 0c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm-2-7h-4v1h4v-1z"
-											/></svg
-										>
-									</button>
+								{#if !order.completed}
+									<div class="flex flex-row gap-2 place-items-center">
+										<button on:click={deleteOrder(order.id)} class="btn btn-square btn-outline">
+											<svg
+												width="24"
+												height="24"
+												xmlns="http://www.w3.org/2000/svg"
+												fill="currentColor"
+												><path
+													d="M19 24h-14c-1.104 0-2-.896-2-2v-17h-1v-2h6v-1.5c0-.827.673-1.5 1.5-1.5h5c.825 0 1.5.671 1.5 1.5v1.5h6v2h-1v17c0 1.104-.896 2-2 2zm0-19h-14v16.5c0 .276.224.5.5.5h13c.276 0 .5-.224.5-.5v-16.5zm-9 4c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm6 0c0-.552-.448-1-1-1s-1 .448-1 1v9c0 .552.448 1 1 1s1-.448 1-1v-9zm-2-7h-4v1h4v-1z"
+												/></svg
+											>
+										</button>
 
-									<h1 class="text-xl">{i + 1}</h1>
-								</div>
+										<h1 class="text-xl">{i + 1}</h1>
+									</div>
+								{/if}
 							</td>
 							<td
 								>{new Date(order.date).toLocaleString([], {
